# Decrease allocations for span attributes