_Before contributing, please read our [contributing guidelines](https://github.com/newrelic/newrelic-ruby-agent/blob/main/CONTRIBUTING.md) and [code of conduct](https://github.com/newrelic/.github/blob/master/CODE_OF_CONDUCT.md)._

# Overview
Previously, our segments were always creating an attributes object in the initializer. Moving the creation of the attributes object to when it is actually being used decreased allocations by 21% in a basic middleware performance test.
Results:
```
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
| name                                       | before    | after     | delta | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 446.70 µs | 425.91 µs | -4.7% |           302 |          236 |       -21.9% |
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
```