@mwlang 2020-08-18
This PR is also introducing the stale issues commenter, which isn't presently on the `main` nor `dev` branches.