@mwlang 2020-08-18
This PR is also seems to be missing the performance benchmark test that was used to validate this change.