@mwlang 2020-08-18
After doing a bit of rebasing between `main` and `dev` and then restoring the performance test's `README.md` this PR is now in tune with `dev` branch so we can safely merge.  However, we're still missing the performance benchmark test was was used to validate this PR's improvements.