@rachelleahklein 2020-08-18
Thanks so very much for sorting out the history here, and sorry for what I added to the confusion. 😅 

Once CI passes, this should be good to go - the performance test you're referring to (results captured in the screenshot aboce) above already existed and is not part of this PR.